+++ lib/al/include/Library/LiveActor/LiveActor.h
@@ -152,23 +152,4 @@ private:
     ActorSceneInfo* mSceneInfo;
     LiveActorFlag* mFlags;
 };
-
-struct LiveActorFlag {
-    bool isDead = true;
-    bool isClipped = false;
-    bool isClippingInvalid = true;
-    bool isDrawClipped = false;
-    bool isCalcAnim = false;
-    bool isModelVisible = false;
-    bool isCollideOff = true;
-    bool field_07 = false;
-    bool isMaterialCodeValid = false;
-    bool isPuddleMaterialValid = false;
-    bool isAreaTargetOn = true;
-    bool isUpdateOn = true;
-
-    LiveActorFlag();
-};
-static_assert(sizeof(LiveActorFlag) == 0xC);
-
 }  // namespace al
