+++ lib/al/src/Library/LiveActor/ActorFlagFunction.cpp
@@ -1,5 +1,6 @@
 #include "Library/LiveActor/ActorFlagFunction.h"
 #include "Library/LiveActor/LiveActor.h"
+#include "Library/LiveActor/LiveActorFlag.h"
 
 namespace al {
 bool isAlive(const LiveActor* actor) {
