+++ lib/al/src/Library/LiveActor/LiveActorFlag.cpp
@@ -1,4 +1,4 @@
-#include "Library/LiveActor/LiveActor.h"
+#include "Library/LiveActor/LiveActorFlag.h"
 
 namespace al {
 LiveActorFlag::LiveActorFlag() = default;
