+++ src/System/ProjectInterface.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace al {
+class SystemKit;
+}
+
+class alProjectInterface {
+    static al::SystemKit* getSystemKit();
+};
